% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_backtest.R
\name{launch_app}
\alias{launch_app}
\title{Launch the QuantLab Shiny Dashboard}
\usage{
launch_app(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{shiny::runApp()}.}
}
\description{
Opens the interactive strategy-testing dashboard in the default browser.
}
