% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{plot.BacktestResult}
\alias{plot.BacktestResult}
\title{Plot the Equity Curve of a BacktestResult}
\usage{
\method{plot}{BacktestResult}(x, ...)
}
\arguments{
\item{x}{A \code{BacktestResult} object.}

\item{...}{Additional arguments passed to \code{plot()}.}
}
\description{
Plot the Equity Curve of a BacktestResult
}
