% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{print.BacktestResult}
\alias{print.BacktestResult}
\title{Print a BacktestResult}
\usage{
\method{print}{BacktestResult}(x, ...)
}
\arguments{
\item{x}{A \code{BacktestResult} object.}

\item{...}{Ignored.}
}
\description{
Print a BacktestResult
}
