% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantlab-package.R
\docType{package}
\name{quantlab-package}
\alias{quantlab}
\alias{quantlab-package}
\title{quantlab: High-Speed Strategy Tester}
\description{
A high-speed backtesting framework for R, utilizing R6 for state management and Rcpp for performance-critical path-dependent simulations. It provides a modular structure for testing simple trading strategies on historical market data, specifically optimized for Polish market assets.
}
\author{
\strong{Maintainer}: Truong Giang Do \email{giangh14cqt@gmail.com}

Authors:
\itemize{
  \item Jan Melan
  \item Sebastian Chmielewski
}

}
\keyword{internal}
