% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{Strategy}
\alias{Strategy}
\title{Strategy Base Class}
\description{
Strategy Base Class

Strategy Base Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Strategy-generate_signals}{\code{Strategy$generate_signals()}}
\item \href{#method-Strategy-clone}{\code{Strategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strategy-generate_signals"></a>}}
\if{latex}{\out{\hypertarget{method-Strategy-generate_signals}{}}}
\subsection{Method \code{generate_signals()}}{
Generate trading signals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strategy$generate_signals(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Market data frame}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer vector of signals (1: Buy, 0: Hold, -1: Sell)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Strategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
