% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{StrategyCrashTest}
\alias{StrategyCrashTest}
\title{Crash Test Strategy}
\description{
Crash Test Strategy

Crash Test Strategy
}
\section{Super class}{
\code{\link[quantlab:Strategy]{quantlab::Strategy}} -> \code{StrategyCrashTest}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{crash_dates}}{Dates to force market entry}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StrategyCrashTest-new}{\code{StrategyCrashTest$new()}}
\item \href{#method-StrategyCrashTest-generate_signals}{\code{StrategyCrashTest$generate_signals()}}
\item \href{#method-StrategyCrashTest-clone}{\code{StrategyCrashTest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrategyCrashTest-new"></a>}}
\if{latex}{\out{\hypertarget{method-StrategyCrashTest-new}{}}}
\subsection{Method \code{new()}}{
Initialize Crash Test Strategy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyCrashTest$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrategyCrashTest-generate_signals"></a>}}
\if{latex}{\out{\hypertarget{method-StrategyCrashTest-generate_signals}{}}}
\subsection{Method \code{generate_signals()}}{
Generate signals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyCrashTest$generate_signals(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Market data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrategyCrashTest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StrategyCrashTest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyCrashTest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
