% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{summary.BacktestResult}
\alias{summary.BacktestResult}
\title{Summarise a BacktestResult}
\usage{
\method{summary}{BacktestResult}(object, ...)
}
\arguments{
\item{object}{A \code{BacktestResult} object.}

\item{...}{Ignored.}
}
\description{
Summarise a BacktestResult
}
