% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loader.R
\name{load_market_data}
\alias{load_market_data}
\title{Load Market Data from CSV}
\usage{
load_market_data(path)
}
\arguments{
\item{path}{Path to the CSV file (Stooq format)}
}
\value{
A data.frame with cleaned market data
}
\description{
Load Market Data from CSV
}
