% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy.R
\name{StrategyATH}
\alias{StrategyATH}
\title{All-Time High Strategy}
\description{
All-Time High Strategy

All-Time High Strategy
}
\section{Super class}{
\code{\link[quantlab:Strategy]{quantlab::Strategy}} -> \code{StrategyATH}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n}}{Rolling window for ATH calculation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-StrategyATH-new}{\code{StrategyATH$new()}}
\item \href{#method-StrategyATH-generate_signals}{\code{StrategyATH$generate_signals()}}
\item \href{#method-StrategyATH-clone}{\code{StrategyATH$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrategyATH-new"></a>}}
\if{latex}{\out{\hypertarget{method-StrategyATH-new}{}}}
\subsection{Method \code{new()}}{
Initialize ATH Strategy
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyATH$new(n = 252)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Rolling window size}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrategyATH-generate_signals"></a>}}
\if{latex}{\out{\hypertarget{method-StrategyATH-generate_signals}{}}}
\subsection{Method \code{generate_signals()}}{
Generate signals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyATH$generate_signals(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Market data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-StrategyATH-clone"></a>}}
\if{latex}{\out{\hypertarget{method-StrategyATH-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StrategyATH$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
